Change API buildDataArray logic to avoid bugs
If the seeded that had no data point for a certain date, the API would build an array only until that point and then return incomplete data. That led to incomplete/cut data being send in API call and resulted in incorrect data being displayed. Could have been observer on Monero with calculating date set before June 5th, 2014.

The source of the problem was if statement comparing if data.date was equal to starting date and then after building an object, adding a day to starting day. With holes in data it was finishing the forEach loop and returning incomplete data.

Fixed by changing the if statement to simply build array when data date is larger or equal to starting date. As a bonus it shortened and simplified the code.